pyproject.toml
[project]
name = "product-validator-search"
version = "0.1.0"
description = "Product Validator Search built with Python and Google ADK integration"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "httpx>=0.27.0",
  "pytrends>=4.9.2",
  "google-adk",
  "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0"
]

[tool.hatch.build.targets.wheel]
packages = ["product_validator_search"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"
